requirements.txt
Flask==1.1.1
gevent==1.4.0
greenlet==0.4.15
httplib2==0.9.2
Jinja2==2.11.1
pusher==2.1.4
requests==2.23.0
requests-unixsocket==0.1.5
simplejson==3.13.2
sseclient-py==1.7
urllib3==1.25.8
virtualenv==15.1.0
Werkzeug==1.0.0
redis==3.4.1
flask-sse